added soc code processing
